moved myBlink folder into the blink folder
