+++ client/src/components/Blinks/MyBlink/Storylist.jsx
@@ -1,6 +1,6 @@
 import React,{useState,useEffect} from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import profileImg from '../../assets/profile.png'
+import profileImg from '../../../assets/profile.png'
 import { currentBlink } from '@/redux/blinkSlice';
 import { IoIosAddCircleOutline } from "react-icons/io";
 import { useNavigate } from 'react-router-dom';
@@ -20,8 +20,8 @@ export default function StoryList({ onSelectStory }) {
     })()
   },[])
   return (
-   <div className='p-1 lg:p-3 md:p-3  xl:p-3 2xl:p-3'>
-    {stories?.length != 0 ?  <div className="flex space-x-4 overflow-x-auto p-4">
+   <div className=''>
+    {stories?.length != 0 ?  <div className="flex space-x-4 overflow-x-auto pl-4 ">
       {stories?.map(story => {
         const getThumbnail = story.stories[0]?.file.replace(/\.mp4/g, '.jpg')
         console.log(getThumbnail);
@@ -36,7 +36,7 @@ export default function StoryList({ onSelectStory }) {
           <p className="text-center text-sm mt-3">{`my Blink`}</p>
         </div>)
       })}
-    </div> : <div className='cursor-pointer' onClick={()=>{navigate('/uploadBlink')}}> 
+    </div> : <div className='cursor-pointer pl-4 ' onClick={()=>{navigate('/uploadBlink')}}> 
           <div className='relative'>
             <img
               className="w-16 h-16 rounded-full border-2  p-1 object-fill"
